[-] warnings

+++ prboom2/src/d_server.c
@@ -409,7 +409,7 @@ int main(int argc, char** argv)
   setupinfo.ticdup = ticdup; setupinfo.extratic = xtratics;
   { /* Random number seed
      * Mirrors the corresponding code in G_ReadOptions */
-    int rngseed = time(NULL);
+    int rngseed = (int)time(NULL);
     setupinfo.game_options[13] = rngseed & 0xff;
     rngseed >>= 8;
     setupinfo.game_options[12] = rngseed & 0xff;
